gltf: add animation and interaction
- animation of robot
- control camera with mouse action
